requirements.txt
picamera2>=0.3.17
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
pyserial>=3.5
requests>=2.31.0
Pillow>=10.0.0